Revert "Refactor client/bot registration to support HostApplicationBuilder"
This reverts commit a89f79e9f38ae27fa65ed6b1cfa4244f7bb65b39.
