+++ src/Disqord.Bot/Hosting/DiscordBotHostBuilderExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
-using System.Threading;
 using Disqord.Bot.Commands.Text;
 using Disqord.DependencyInjection.Extensions;
 using Disqord.Hosting;
@@ -33,34 +33,16 @@ public static class DiscordBotHostBuilderExtensions
             var discordContext = new DiscordBotHostingContext();
             configure?.Invoke(context, discordContext);
 
-            services.AddDiscordBot<TDiscordBot, TDiscordBotConfiguration>(discordContext);
+            services.AddDiscordBot<TDiscordBot>();
+            services.ConfigureDiscordBot<TDiscordBotConfiguration>(context, discordContext);
+            services.ConfigureDiscordClient(context, discordContext);
         });
 
         return builder;
     }
 
-    public static IServiceCollection AddDiscordBot(this IServiceCollection services, DiscordBotHostingContext context)
-    {
-        return services.AddDiscordBot<DiscordBot>(context);
-    }
-
-    public static IServiceCollection AddDiscordBot<TDiscordBot>(this IServiceCollection services, DiscordBotHostingContext context)
-        where TDiscordBot : DiscordBot
-    {
-        return services.AddDiscordBot<TDiscordBot, DiscordBotConfiguration>(context);
-    }
-
-    public static IServiceCollection AddDiscordBot<TDiscordBot, TDiscordBotConfiguration>(this IServiceCollection services, DiscordBotHostingContext context)
-        where TDiscordBot : DiscordBot
-        where TDiscordBotConfiguration : DiscordBotBaseConfiguration, new()
-    {
-        services.AddDiscordBot<TDiscordBot>();
-        services.ConfigureDiscordBot<TDiscordBotConfiguration>(context);
-        services.ConfigureDiscordClient(context);
-        return services;
-    }
-
-    internal static void ConfigureDiscordBot<TBotConfiguration>(this IServiceCollection services, DiscordBotHostingContext discordContext)
+    [EditorBrowsable(EditorBrowsableState.Never)]
+    public static void ConfigureDiscordBot<TBotConfiguration>(this IServiceCollection services, HostBuilderContext context, DiscordBotHostingContext discordContext)
         where TBotConfiguration : DiscordBotBaseConfiguration, new()
     {
         if (discordContext.OwnerIds != null || discordContext.ApplicationId != null)
@@ -92,45 +74,5 @@ public static class DiscordBotHostBuilderExtensions
         services.AddHostedService<DiscordBotSetupService>();
 
         services.TryAddSingleton<DiscordBotMasterService>();
-    }
-
-    internal static IServiceCollection AddDiscordBot<TDiscordBot>(this IServiceCollection services)
-        where TDiscordBot : DiscordBot
-    {
-        services.AddDiscordClient();
-
-        if (services.TryAddSingleton<TDiscordBot>())
-        {
-            var callCount = 0;
-
-            TDiscordBot GetBotService(IServiceProvider services)
-            {
-                if (Interlocked.Increment(ref callCount) > 1)
-                    throw new InvalidOperationException($"Disqord detected a circular dependency for the bot client of type '{typeof(TDiscordBot)}'. "
-                        + "This means that most likely your prefix provider or another service depends on the bot client and vice versa.");
-
-                var service = services.GetRequiredService<TDiscordBot>();
-                Interlocked.Decrement(ref callCount);
-                return service;
-            }
-
-            if (typeof(TDiscordBot) != typeof(DiscordBot))
-                services.TryAddSingleton<DiscordBot>(GetBotService);
-
-            services.TryAddSingleton<DiscordBotBase>(GetBotService);
-            services.Replace(ServiceDescriptor.Singleton<DiscordClientBase>(GetBotService));
-        }
-
-        services.AddDiscordBotDependencies();
-        return services;
-    }
-
-    internal static void AddDiscordBotDependencies(this IServiceCollection services)
-    {
-        services.AddPrefixProvider();
-        services.AddCommands();
-        services.AddCommandContextAccessor();
-        services.AddApplicationCommandLocalizer();
-        services.AddApplicationCommandCacheProvider();
     }
 }
